requirements.txt
Flask==2.1.3
numpy==1.21.6
Pillow==7.1.2
torch==1.12.0
git+https://github.com/openai/CLIP.git